requirements.txt
tk
pycryptodome